Extract the Navigator interfaces in their own file and document them
